Add customer card page and routing
